LayoutTests/http/wpt/mediarecorder/pause-recording-2.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <script>
function waitFor(duration)
{
    return new Promise((resolve) => setTimeout(resolve, duration));
}

function waitForEvent(object, eventName, testName)
{
    return new Promise((resolve, reject) => {
        object.addEventListener(eventName, resolve, { once: true });
        setTimeout(() => reject("waitForEvent " + (testName ? (testName + " ") : "") + "timed out for " + eventName), 2000);
    });
}

promise_test(async (test) => {
    const stream = await navigator.mediaDevices.getUserMedia({ audio: true, video: true });

    const recorder = new MediaRecorder(stream, { mimeType: "video/webm" });
    const dataPromise = new Promise(resolve => recorder.ondataavailable = (e) => resolve(e.data));

    const startPromise = new Promise(resolve => recorder.onstart = resolve);
    recorder.start();
    await startPromise;

    setTimeout(() => recorder.pause(), 50);
    setTimeout(() => recorder.resume(), 950);

    await waitFor(2000);
    recorder.stop();
    const blob = await dataPromise;

    const video = document.createElement("video");
    video.disableRemotePlayback = true;
    const metadataPromise = waitForEvent(video, "loadedmetadata");

    // webm generated don't have a duration located in the metadata. Determine the duration with MediaSource.
    if (!window.ManagedMediaSource?.isTypeSupported(blob.type)) {
        const url = URL.createObjectURL(blob);
        video.src = url;
        await metadataPromise;
        URL.revokeObjectURL(url);
        return;
    }

    const source = new ManagedMediaSource();
    video.srcObject = source;
    await waitForEvent(source, "sourceopen");
    const buffer = source.addSourceBuffer(blob.type);
    buffer.appendBuffer(await blob.arrayBuffer());
    await waitForEvent(buffer, "update");
    await metadataPromise;
    source.endOfStream();
    await waitForEvent(source, "sourceended");
    assert_less_than(video.duration, 1.5);
}, "Pausing and resuming the WebM recording should impact the video duration");
    </script>
</body>
</html>
